#1686 - Start Refactor 3, change storage type to long (bigint), use jsonb for most stat data, store game data in tables. Change some id's to snowflakes.

+++ core/src/main/java/me/mykindos/betterpvp/core/client/achievements/impl/game/wins/ChampionsWinAchievement.java
@@ -29,7 +29,7 @@ public class ChampionsWinAchievement extends SingleSimpleAchievement {
         super("Champions Wins", key,
                 AchievementCategories.GAME_CHAMPIONS_WINS,
                 AchievementType.GLOBAL,
-                (double) goal,
+                (long) goal,
                 GameTeamMapNativeStat.builder()
                         .action(GameTeamMapNativeStat.Action.WIN)
                         .build()
